feat: adjusted sql similarity function to leverage query type + oso models used
